Binh

+++ src/BirdStoreAndroid/app/src/main/java/com/example/birdstoreandroid/Model/CreateOrderResponse.java
@@ -156,7 +156,7 @@ public class CreateOrderResponse {
         private Product product;
         private String order_id;
         private Object order;
-        private int price;
+        private float price;
         private int quantity;
         private String user_id;
         private Object user;
@@ -197,11 +197,11 @@ public class CreateOrderResponse {
             this.order = order;
         }
 
-        public int getPrice() {
+        public float getPrice() {
             return price;
         }
 
-        public void setPrice(int price) {
+        public void setPrice(float price) {
             this.price = price;
         }
 
@@ -265,7 +265,7 @@ public class CreateOrderResponse {
     public static class Product {
         private String category_id;
         private String image;
-        private int price;
+        private float price;
         private String name;
         private boolean sex;
         private String description;
@@ -296,11 +296,11 @@ public class CreateOrderResponse {
             this.image = image;
         }
 
-        public int getPrice() {
+        public float getPrice() {
             return price;
         }
 
-        public void setPrice(int price) {
+        public void setPrice(float price) {
             this.price = price;
         }
